starting to move things into files, broke something already
